HC-3637 Improve responses & fix validator (#23)
* HC-3637 After-design KB backend improvements

* Review fixes

+++ src/Controller/HelpcrunchController.php
@@ -3,8 +3,10 @@
 namespace Helpcrunch\Controller;
 
 use Helpcrunch\Entity\HelpcrunchEntity;
+use Helpcrunch\Helper\ParametersValidatorHelper;
 use Helpcrunch\Repository\HelpcrunchRepository;
 use Helpcrunch\Response\EntitiesBatchResponse;
+use Helpcrunch\Response\EntityNotFoundResponse;
 use Helpcrunch\Response\EntityResponse;
 use Helpcrunch\Response\ErrorResponse;
 use Helpcrunch\Response\InnerErrorCodes;
@@ -14,13 +16,13 @@ use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\Routing\ClassResourceInterface;
+use FOS\RestBundle\Controller\Annotations as Rest;
 use Helpcrunch\Traits\HelpcrunchServicesTrait;
 use Helpcrunch\Validator\Validator;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArrayInput;
@@ -75,9 +77,20 @@ abstract class HelpcrunchController extends FOSRestController implements ClassRe
         return new EntitiesBatchResponse($this->getRepository()->findEntities($offset, $limit));
     }
 
-    public function getAction(int $id)
+    /**
+     * @param int $id
+     * @return JsonResponse
+     */
+    public function getAction($id): JsonResponse
     {
-        return new EntityResponse($this->findEntityById($id));
+        if (!ParametersValidatorHelper::isValidId($id)) {
+            return new ErrorResponse('Invalid ID', InnerErrorCodes::INVALID_ENTITY_ID);
+        }
+        if (!($entity = $this->getRepository()->find($id))) {
+            return new EntityNotFoundResponse(self::$entityClassName);
+        }
+
+        return new EntityResponse($entity);
     }
 
     public function postAction(Request $request): JsonResponse
@@ -96,12 +109,24 @@ abstract class HelpcrunchController extends FOSRestController implements ClassRe
         $this->entityManager->persist($entity);
         $this->entityManager->flush();
 
-        return new EntityResponse($this->findEntityById($entity->id), 'entity created', Response::HTTP_CREATED);
+        return new EntityResponse($this->getRepository()->find($entity->id), 'entity created', Response::HTTP_CREATED);
     }
 
-    public function putAction(Request $request, int $id): JsonResponse
+    /**
+     * @param Request $request
+     * @param int $id
+     * @return JsonResponse
+     * @throws \Doctrine\Common\Annotations\AnnotationException
+     * @throws \ReflectionException
+     */
+    public function putAction(Request $request, $id): JsonResponse
     {
-        $entity = $this->findEntityById($id);
+        if (!ParametersValidatorHelper::isValidId($id)) {
+            return new ErrorResponse('Invalid ID', InnerErrorCodes::INVALID_ENTITY_ID);
+        }
+        if (!($entity = $this->getRepository()->find($id))) {
+            return new EntityNotFoundResponse(self::$entityClassName);
+        }
 
         $validator = new Validator($this->container);
         if (!($entity = $validator->isValid($entity, $request->request->all()))) {
@@ -116,9 +141,18 @@ abstract class HelpcrunchController extends FOSRestController implements ClassRe
         return new EntityResponse($entity, 'entity updated');
     }
 
-    public function deleteAction(int $id): JsonResponse
+    /**
+     * @param int $id
+     * @return JsonResponse
+     */
+    public function deleteAction($id): JsonResponse
     {
-        $entity = $this->findEntityById($id);
+        if (!ParametersValidatorHelper::isValidId($id)) {
+            return new ErrorResponse('Invalid ID', InnerErrorCodes::INVALID_ENTITY_ID);
+        }
+        if (!($entity = $this->getRepository()->find($id))) {
+            return new EntityNotFoundResponse(self::$entityClassName);
+        }
 
         $this->entityManager->remove($entity);
         $this->entityManager->flush();
@@ -132,16 +166,6 @@ abstract class HelpcrunchController extends FOSRestController implements ClassRe
     protected function getRepository()
     {
         return $this->getDoctrine()->getRepository(static::$entityClassName);
-    }
-
-    protected function findEntityById(int $id): HelpcrunchEntity
-    {
-        $entity = $this->getRepository()->find($id);
-        if (!$entity) {
-            throw new NotFoundHttpException();
-        }
-
-        return $entity;
     }
 
     protected function getNewEntity(): HelpcrunchEntity
